Rename constants file
