+++ lib/daru/view/constants_data_table.rb
